切换角色bug修改，技能伤害命中处理

+++ Src/Lib/Common/Data/SkillDefine.cs
@@ -23,7 +23,7 @@ namespace Common.Data
         public bool Bullet { get; set; }
         public float BulletSpeed { get; set; }
         public string BulletResource { get; set; }
-        public float AOERange { get; set; }
+        public int AOERange { get; set; }
         public float Duration { get; set; }
         public float Interval { get; set; }
         public List<float> HitTimes { get; set; }
